Restructure files to group Express functionality under same namespace
